Allow preemption on UVC CCI attribute gets to ensure smooth streaming
This change dispatches attribute get requests in a pt coroutine. It
re-implements some FLIR SDK I2C functionality to place a bunch of
yield statements in there, so that we don't starve the USB pipe
and get out of sync.

Note that this change just affects attribute get requests, sets are
still performed synchronously with the USB control interrupt handling.