Don't acknowledge queued CCI attribute sets until complete
In the last commit, the attribute sets were acknowledged and
queued, but there was no way to indicate a failure if something
happened, nor was there a standard way for the host to attempt to
block on those requests.

This works the same way as attribute gets now. Attribute sets are
simply not acknowledged until they are complete, effectively
making them synchronous.